+++ __mocks__/glob.ts
@@ -1,0 +1,5 @@
+import * as mockTestFileData from "./testFileData.mock";
+
+const sync = jest.fn(() => Object.keys(mockTestFileData.mock));
+
+export { sync };