+++ __tests__/core/test-server/test-server-factory.tests.ts
@@ -1,5 +1,5 @@
 import { KarmaServer } from "../../../src/core/karma/karma-server";
-import { ProjectType } from "../../../src/model/project-type.enum";
+import { ProjectType } from "../../../src/model/enums/project-type.enum";
 import { AngularServer } from "../../../src/core/angular/angular-server";
 import { TestServerFactory } from "../../../src/core/test-server/test-server-factory";
 import { KarmaEventListener } from "../../../src/core/integration/karma-event-listener";