+++ src/core/shared/event-emitter.ts
@@ -1,10 +1,14 @@
 import { KarmaEvent } from "./../../model/karma-event";
-import { TestState } from "./../../model/test-state.enum";
-import { TestEvent } from "vscode-test-adapter-api";
+import { TestState } from "../../model/enums/test-state.enum";
+import { TestEvent, TestRunStartedEvent, TestRunFinishedEvent, TestSuiteEvent, TestDecoration } from "vscode-test-adapter-api";
 import { TestResultToTestStateMapper } from "../test-explorer/test-result-to-test-state.mapper";
+import * as vscode from "vscode";
+import { SpecCompleteResponse } from "../../model/spec-complete-response";
 
 export class EventEmitter {
-  public constructor(private readonly eventEmitterInterface: any) {}
+  public constructor(
+    private readonly eventEmitterInterface: vscode.EventEmitter<TestRunStartedEvent | TestRunFinishedEvent | TestSuiteEvent | TestEvent>
+  ) {}
 
   public emitTestStateEvent(testName: string, testState: TestState) {
     const testEvent = { type: "test", test: testName, state: testState } as TestEvent;
@@ -18,9 +22,58 @@ export class EventEmitter {
     const testEvent = { type: "test", test: testName, state: testState } as TestEvent;
 
     if (karmaEvent.results.failureMessages.length > 0) {
-      testEvent.message = karmaEvent.results.failureMessages[0];
+      testEvent.decorations = this.createDecorations(karmaEvent.results);
+      testEvent.message = this.createErrorMessage(karmaEvent.results);
     }
 
     this.eventEmitterInterface.fire(testEvent);
+  }
+
+  private createErrorMessage(results: SpecCompleteResponse): string {
+    const failureMessage = results.failureMessages[0];
+    const message = failureMessage.split("\n")[0];
+
+    if (!results.filePath) {
+      return message;
+    }
+
+    try {
+      const errorLineAndColumnCollection = failureMessage.substring(failureMessage.indexOf(results.filePath)).split(":");
+      const lineNumber = parseInt(errorLineAndColumnCollection[1], undefined);
+      const columnNumber = parseInt(errorLineAndColumnCollection[2], undefined);
+
+      if (isNaN(lineNumber) || isNaN(columnNumber)) {
+        return message;
+      }
+
+      return `${message} (line:${lineNumber} column:${columnNumber})`;
+    } catch (error) {
+      return failureMessage;
+    }
+  }
+
+  private createDecorations(results: SpecCompleteResponse): TestDecoration[] | undefined {
+    if (!results.filePath) {
+      return undefined;
+    }
+
+    try {
+      const decorations = results.failureMessages.map((failureMessage: string) => {
+        const errorLineAndColumnCollection = failureMessage.substring(failureMessage.indexOf(results.filePath as string)).split(":");
+        const lineNumber = parseInt(errorLineAndColumnCollection[1], undefined);
+        return {
+          line: lineNumber,
+          message: failureMessage.split("\n")[0],
+        };
+      });
+
+      if (decorations.some(x => isNaN(x.line))) {
+        return undefined;
+      }
+
+      return decorations;
+    } catch (error) {
+      return undefined;
+    }
   }
 }